requirements.txt
mcp>=0.1.0
fastmcp>=0.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.25.0
httpx>=0.23.0
pillow>=9.0.0
numpy>=1.20.0
